code quality

+++ JRomManager/src/jrm/batch/Compressor.java
@@ -11,6 +11,7 @@ import java.nio.file.StandardCopyOption;
 import java.nio.file.attribute.BasicFileAttributes;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
@@ -40,6 +41,11 @@ import lombok.Getter;
 
 public class Compressor implements HTMLRenderer
 {
+	private static final String OK = "OK";
+	private static final String FAILED = "Failed";
+	private static final String CRUNCHING = "Crunching ";
+	private static final String EXTRACTING = "extracting ";
+	private static final String PROCESSING = "Processing ";
 	private final Session session;
 	private final AtomicInteger cnt;
 	private final int total;
@@ -114,145 +120,108 @@ public class Compressor implements HTMLRenderer
 	{
 		try
 		{
-			cb.apply("Processing "+file.getName());
+			cb.apply(PROCESSING+file.getName());
 			final Path tmpfile = IOUtils.createTempFile("JRM", ".7z");
 			Files.delete(tmpfile);
 			final var newfile = new File(file.getParentFile(),FilenameUtils.getBaseName(file.getName())+".7z");
-			try(final var archive = new SevenZipArchive(session, file, true, new ProgressNarchiveCallBack(progress)))
+			if (sevenZip2SevenZip(file, cb, tmpfile, newfile) && Files.exists(tmpfile))
 			{
-				progress.setProgress(toHTML("extracting " + toItalic(StringEscapeUtils.escapeHtml4(file.getName()))), cnt.get(), total);
-				if(archive.extract()==0)
-				{
-					try(final var newarchive = new SevenZipArchive(session, tmpfile.toFile(), new ProgressNarchiveCallBack(progress)))
-					{
-						final var basedir = archive.getTempDir().toPath();
-						Files.walkFileTree(basedir, new SimpleFileVisitor<Path>() {
-							@Override
-							public FileVisitResult visitFile(Path file, BasicFileAttributes attr) throws IOException
-							{
-								newarchive.addStdIn(Files.newInputStream(file), basedir.relativize(file).toString());
-								return FileVisitResult.CONTINUE;
-							}
-							
-							@Override
-							public FileVisitResult postVisitDirectory(Path dir, IOException exc) throws IOException
-							{
-								return FileVisitResult.CONTINUE;
-							}
-						});
-						progress.setProgress(toHTML("Crunching " + toItalic(StringEscapeUtils.escapeHtml4(newfile.getName()))), cnt.get(), total);
-					}
-				}
-				else
-				{
-					Files.deleteIfExists(tmpfile);
-					cb.apply("extract failed");
-					return null;
-				}
-			}
-			catch(Exception e)
-			{
-				Files.deleteIfExists(tmpfile);
-				cb.apply("7z creation failed");
-				return null;
-			}
-			if(Files.exists(tmpfile))
-			{
-				if(FileUtils.deleteQuietly(file))
+				if (FileUtils.deleteQuietly(file))
 				{
 					FileUtils.moveFile(tmpfile.toFile(), newfile);
 					scb.apply(newfile);
-					cb.apply("OK");
+					cb.apply(OK);
 					return newfile;
 				}
 				else
-					cb.apply("Failed");
+					cb.apply(FAILED);
 			}
 		}
 		catch(IOException e)
 		{
-			cb.apply("Failed");
+			cb.apply(FAILED);
 		}
 		finally
 		{
-//			progress.setProgress("", null, null, "");
+			// do nothing
 		}
 		return null;
+	}
+
+	/**
+	 * @param file
+	 * @param cb
+	 * @param tmpfile
+	 * @param newfile
+	 * @throws IOException
+	 */
+	private boolean sevenZip2SevenZip(final File file, final UpdResultCallBack cb, final Path tmpfile, final File newfile) throws IOException
+	{
+		try(final var archive = new SevenZipArchive(session, file, true, new ProgressNarchiveCallBack(progress)))
+		{
+			progress.setProgress(toHTML(EXTRACTING + toItalic(StringEscapeUtils.escapeHtml4(file.getName()))), cnt.get(), total);
+			if(archive.extract()==0)
+			{
+				try(final var newarchive = new SevenZipArchive(session, tmpfile.toFile(), new ProgressNarchiveCallBack(progress)))
+				{
+					final var basedir = archive.getTempDir().toPath();
+					Files.walkFileTree(basedir, new SimpleFileVisitor<Path>() {
+						@Override
+						public FileVisitResult visitFile(Path file, BasicFileAttributes attr) throws IOException
+						{
+							newarchive.addStdIn(Files.newInputStream(file), basedir.relativize(file).toString());
+							return FileVisitResult.CONTINUE;
+						}
+						
+						@Override
+						public FileVisitResult postVisitDirectory(Path dir, IOException exc) throws IOException
+						{
+							return FileVisitResult.CONTINUE;
+						}
+					});
+					progress.setProgress(toHTML(CRUNCHING + toItalic(StringEscapeUtils.escapeHtml4(newfile.getName()))), cnt.get(), total);
+				}
+			}
+			else
+			{
+				Files.deleteIfExists(tmpfile);
+				cb.apply("extract failed");
+				return false;
+			}
+		}
+		catch(Exception e)
+		{
+			Files.deleteIfExists(tmpfile);
+			cb.apply("7z creation failed");
+			return false;
+		}
+		return true;
 	}
 
 	public File sevenZip2Zip(final File file, final boolean tzip, final UpdResultCallBack cb, final UpdSrcCallBack scb)
 	{
 		try
 		{
-			cb.apply("Processing "+file.getName());
+			cb.apply(PROCESSING+file.getName());
 			final var tmpfile = IOUtils.createTempFile("JRM", ".zip");
 			Files.delete(tmpfile);
 			final var newfile = new File(file.getParentFile(),FilenameUtils.getBaseName(file.getName())+".zip");
-			try(final var archive = new SevenZipArchive(session, file, false, new ProgressNarchiveCallBack(progress)))
-			{
-				progress.setProgress(toHTML("extracting " + toItalic(StringEscapeUtils.escapeHtml4(file.getName()))), cnt.get(), total);
-				if(archive.extract()==0)
-				{
-					final File basedir = archive.getTempDir();
-					final Map<String, Object> env = new HashMap<>();
-					env.put("create", "true"); //$NON-NLS-1$ //$NON-NLS-2$
-					env.put("useTempFile", FileUtils.sizeOf(basedir) > ZipTempThreshold.valueOf(session.getUser().getSettings().getProperty(jrm.misc.SettingsEnum.zip_temp_threshold, ZipTempThreshold._10MB.toString())).getThreshold()); //$NON-NLS-1$ //$NON-NLS-2$
-					env.put("compressionLevel", tzip ? 1 :  ZipLevel.valueOf(session.getUser().getSettings().getProperty(jrm.misc.SettingsEnum.zip_compression_level, ZipLevel.DEFAULT.toString())).getLevel()); //$NON-NLS-1$ //$NON-NLS-2$
-					FileUtils.forceMkdirParent(tmpfile.toFile());
-					progress.setProgress(toHTML("creating " + toItalic(StringEscapeUtils.escapeHtml4(newfile.getName()))), cnt.get(), total);
-					try(final var dstarchive = new ZipArchive(session, tmpfile.toFile(), new ProgressNarchiveCallBack(progress)))
-					{
-						dstarchive.compress_custom(new CustomVisitor(basedir.toPath()) {
-							@Override
-							public FileVisitResult visitFile(Path file, BasicFileAttributes attr) throws IOException
-							{
-								final var dst = getFileSystem().getPath(basedir.toPath().relativize(file).toString());
-								if(dst.getParent()!=null)
-									Files.createDirectories(dst.getParent());
-								Files.copy(file, dst, StandardCopyOption.REPLACE_EXISTING);
-								return FileVisitResult.CONTINUE;
-							}
-							
-							@Override
-							public FileVisitResult postVisitDirectory(Path dir, IOException exc) throws IOException
-							{
-								final var dst = getFileSystem().getPath(basedir.toPath().relativize(dir).toString());
-								Files.createDirectories(dst);
-								return FileVisitResult.CONTINUE;
-							}
-						}, env);
-					}
-				}
-				else
-				{
-					Files.deleteIfExists(tmpfile);
-					cb.apply("extract failed");
-					return null;
-				}
-			}
-			catch (Exception e)
-			{
-				Log.err(e.getMessage(),e);
-				Files.deleteIfExists(tmpfile);
-				cb.apply("zip creation failed");
-				return null;
-			}
-			if(Files.exists(tmpfile))
+			if(sevenZip2Zip(file, tzip, cb, tmpfile, newfile)&&Files.exists(tmpfile))
 			{
 				if(FileUtils.deleteQuietly(file))
 				{
 					FileUtils.moveFile(tmpfile.toFile(), newfile);
 					scb.apply(newfile);
-					cb.apply("OK");
+					cb.apply(OK);
 					return newfile;
 				}
 				else
-					cb.apply("Failed");
+					cb.apply(FAILED);
 			}
 		}
 		catch (final IOException e)
 		{
-			cb.apply("Failed");
+			cb.apply(FAILED);
 		}
 		finally
 		{
@@ -261,11 +230,73 @@ public class Compressor implements HTMLRenderer
 		return null;
 	}
 
+	/**
+	 * @param file
+	 * @param tzip
+	 * @param cb
+	 * @param tmpfile
+	 * @param newfile
+	 * @throws IOException
+	 */
+	private boolean sevenZip2Zip(final File file, final boolean tzip, final UpdResultCallBack cb, final Path tmpfile, final File newfile) throws IOException
+	{
+		try(final var archive = new SevenZipArchive(session, file, false, new ProgressNarchiveCallBack(progress)))
+		{
+			progress.setProgress(toHTML(EXTRACTING + toItalic(StringEscapeUtils.escapeHtml4(file.getName()))), cnt.get(), total);
+			if(archive.extract()==0)
+			{
+				final File basedir = archive.getTempDir();
+				final Map<String, Object> env = new HashMap<>();
+				env.put("create", "true"); //$NON-NLS-1$ //$NON-NLS-2$
+				env.put("useTempFile", FileUtils.sizeOf(basedir) > ZipTempThreshold.valueOf(session.getUser().getSettings().getProperty(jrm.misc.SettingsEnum.zip_temp_threshold, ZipTempThreshold._10MB.toString())).getThreshold()); //$NON-NLS-1$ //$NON-NLS-2$
+				env.put("compressionLevel", tzip ? 1 :  ZipLevel.valueOf(session.getUser().getSettings().getProperty(jrm.misc.SettingsEnum.zip_compression_level, ZipLevel.DEFAULT.toString())).getLevel()); //$NON-NLS-1$ //$NON-NLS-2$
+				FileUtils.forceMkdirParent(tmpfile.toFile());
+				progress.setProgress(toHTML("creating " + toItalic(StringEscapeUtils.escapeHtml4(newfile.getName()))), cnt.get(), total);
+				try(final var dstarchive = new ZipArchive(session, tmpfile.toFile(), new ProgressNarchiveCallBack(progress)))
+				{
+					dstarchive.compress_custom(new CustomVisitor(basedir.toPath()) {
+						@Override
+						public FileVisitResult visitFile(Path file, BasicFileAttributes attr) throws IOException
+						{
+							final var dst = getFileSystem().getPath(basedir.toPath().relativize(file).toString());
+							if(dst.getParent()!=null)
+								Files.createDirectories(dst.getParent());
+							Files.copy(file, dst, StandardCopyOption.REPLACE_EXISTING);
+							return FileVisitResult.CONTINUE;
+						}
+						
+						@Override
+						public FileVisitResult postVisitDirectory(Path dir, IOException exc) throws IOException
+						{
+							final var dst = getFileSystem().getPath(basedir.toPath().relativize(dir).toString());
+							Files.createDirectories(dst);
+							return FileVisitResult.CONTINUE;
+						}
+					}, env);
+				}
+			}
+			else
+			{
+				Files.deleteIfExists(tmpfile);
+				cb.apply("extract failed");
+				return false;
+			}
+		}
+		catch (Exception e)
+		{
+			Log.err(e.getMessage(),e);
+			Files.deleteIfExists(tmpfile);
+			cb.apply("zip creation failed");
+			return false;
+		}
+		return true;
+	}
+
 	public File zip2Zip(final File file, final UpdResultCallBack cb, final UpdSrcCallBack scb)
 	{
 		try
 		{
-			cb.apply("Processing "+file.getName());
+			cb.apply(PROCESSING+file.getName());
 			final var tmpfile = IOUtils.createTempFile("JRM", ".zip");
 			Files.delete(tmpfile);
 			final var newfile = new File(file.getParentFile(),FilenameUtils.getBaseName(file.getName())+".zip");
@@ -276,7 +307,7 @@ public class Compressor implements HTMLRenderer
 				env.put("create", "true"); //$NON-NLS-1$ //$NON-NLS-2$
 				env.put("useTempFile", size(basedir) > ZipTempThreshold.valueOf(session.getUser().getSettings().getProperty(jrm.misc.SettingsEnum.zip_temp_threshold, ZipTempThreshold._10MB.toString())).getThreshold()); //$NON-NLS-1$ //$NON-NLS-2$
 				env.put("compressionLevel", ZipLevel.valueOf(session.getUser().getSettings().getProperty(jrm.misc.SettingsEnum.zip_compression_level, ZipLevel.DEFAULT.toString())).getLevel()); //$NON-NLS-1$
-				progress.setProgress(toHTML("Crunching " + toItalic(StringEscapeUtils.escapeHtml4(newfile.getName()))), cnt.get(), total);
+				progress.setProgress(toHTML(CRUNCHING + toItalic(StringEscapeUtils.escapeHtml4(newfile.getName()))), cnt.get(), total);
 				try (final var newarchive = new ZipArchive(session, tmpfile.toFile(), new ProgressNarchiveCallBack(progress)))
 				{
 					newarchive.compress_custom(new CustomVisitor(basedir) {
@@ -306,16 +337,16 @@ public class Compressor implements HTMLRenderer
 				{
 					FileUtils.moveFile(tmpfile.toFile(), newfile);
 					scb.apply(newfile);
-					cb.apply("OK");
+					cb.apply(OK);
 					return newfile;
 				}
 				else
-					cb.apply("Failed");
+					cb.apply(FAILED);
 			}
 		}
 		catch(IOException e)
 		{
-			cb.apply("failed");
+			cb.apply(FAILED);
 		}
 		finally
 		{
@@ -328,40 +359,17 @@ public class Compressor implements HTMLRenderer
 	{
 		try
 		{
-			cb.apply("Processing "+file.getName());
+			cb.apply(PROCESSING + file.getName());
 			final var tmpfile = IOUtils.createTempFile("JRM", ".7z");
 			Files.delete(tmpfile);
 			final var newfile = new File(file.getParentFile(),FilenameUtils.getBaseName(file.getName())+".7z");
-			try(final var archive = new SevenZipArchive(session, tmpfile.toFile(), new ProgressNarchiveCallBack(progress)))
-			{
-				progress.setProgress(toHTML("extracting " + toItalic(StringEscapeUtils.escapeHtml4(file.getName()))), cnt.get(), total);
-				try(final var srcarchive = new ZipArchive(session, file, true, new ProgressNarchiveCallBack(progress));)
-				{
-					srcarchive.extract_custom(new CustomVisitor() {
-						@Override
-						public FileVisitResult visitFile(Path file, BasicFileAttributes attr) throws IOException
-						{
-							archive.addStdIn(Files.newInputStream(file), getSourcePath().relativize(file).toString());
-							return FileVisitResult.CONTINUE;
-						}
-					});
-				}
-				progress.setProgress(toHTML("Crunching " + toItalic(StringEscapeUtils.escapeHtml4(newfile.getName()))), cnt.get(), total);
-			}
-			catch (Exception e)
-			{
-				Log.err(e.getMessage(),e);
-				Files.deleteIfExists(tmpfile);
-				cb.apply("7z creation failed");
-				return null;
-			}
-			if(Files.exists(tmpfile))
+			if(zip2SevenZip(file, cb, tmpfile, newfile)&&Files.exists(tmpfile))
 			{
 				if(FileUtils.deleteQuietly(file))
 				{
 					FileUtils.moveFile(tmpfile.toFile(), newfile);
 					scb.apply(newfile);
-					cb.apply("OK");
+					cb.apply(OK);
 					return newfile;
 				}
 				else
@@ -379,16 +387,51 @@ public class Compressor implements HTMLRenderer
 		return null;
 	}
 
+	/**
+	 * @param file
+	 * @param cb
+	 * @param tmpfile
+	 * @param newfile
+	 * @throws IOException
+	 */
+	private boolean zip2SevenZip(final File file, final UpdResultCallBack cb, final Path tmpfile, final File newfile) throws IOException
+	{
+		try(final var archive = new SevenZipArchive(session, tmpfile.toFile(), new ProgressNarchiveCallBack(progress)))
+		{
+			progress.setProgress(toHTML(EXTRACTING + toItalic(StringEscapeUtils.escapeHtml4(file.getName()))), cnt.get(), total);
+			try(final var srcarchive = new ZipArchive(session, file, true, new ProgressNarchiveCallBack(progress));)
+			{
+				srcarchive.extract_custom(new CustomVisitor() {
+					@Override
+					public FileVisitResult visitFile(Path file, BasicFileAttributes attr) throws IOException
+					{
+						archive.addStdIn(Files.newInputStream(file), getSourcePath().relativize(file).toString());
+						return FileVisitResult.CONTINUE;
+					}
+				});
+			}
+			progress.setProgress(toHTML(CRUNCHING + toItalic(StringEscapeUtils.escapeHtml4(newfile.getName()))), cnt.get(), total);
+		}
+		catch (Exception e)
+		{
+			Log.err(e.getMessage(),e);
+			Files.deleteIfExists(tmpfile);
+			cb.apply("7z creation failed");
+			return false;
+		}
+		return true;
+	}
+
 	public File zip2TZip(final File file, final boolean force, final UpdResultCallBack cb)
 	{
 		try
 		{
 			progress.setProgress(toHTML("TorrentZipping " + toItalic(StringEscapeUtils.escapeHtml4(file.getName()))), cnt.get(), total);
-			cb.apply("Processing "+file.getName());
+			cb.apply(PROCESSING+file.getName());
 			final Set<TrrntZipStatus> status = new TorrentZip(new ProgressTZipCallBack(progress), new SimpleTorrentZipOptions(force,false)).Process(file);
 			if(status.contains(TrrntZipStatus.ValidTrrntzip))
 			{
-				cb.apply("OK");
+				cb.apply(OK);
 				return file;
 			}
 			cb.apply(status.toString());
@@ -410,54 +453,80 @@ public class Compressor implements HTMLRenderer
 		{
 			case SEVENZIP:
 			{
-				switch (FilenameUtils.getExtension(file.getName()))
-				{
-					case "zip":
-						zip2SevenZip(file, cb, scb);
-						break;
-					case "7z":
-						if (force)
-							sevenZip2SevenZip(file, cb, scb);
-						else
-							cb.apply("Skipped");
-						break;
-					default:
-						sevenZip2SevenZip(file, cb, scb);
-						break;
-				}
+				compressToSevenZip(file, force, cb, scb);
 				break;
 			}
 			case ZIP:
 			{
-				switch (FilenameUtils.getExtension(file.getName()))
-				{
-					case "zip":
-						if (force)
-							zip2Zip(file, cb, scb);
-						else
-							cb.apply("Skipped");
-						break;
-					default:
-						sevenZip2Zip(file, false, cb, scb);
-						break;
-				}
+				compressToZip(file, force, cb, scb);
 				break;
 			}
 			case TZIP:
 			{
-				switch (FilenameUtils.getExtension(file.getName()))
-				{
-					case "zip":
-						zip2TZip(file, force, cb);
-						break;
-					default:
-						final File f = sevenZip2Zip(file, true, cb, scb);
-						if (f != null && f.exists())
-							zip2TZip(f, force, cb);
-						break;
-				}
+				compressToTZip(file, force, cb, scb);
 				break;
 			}
+		}
+	}
+
+	/**
+	 * @param file
+	 * @param force
+	 * @param cb
+	 * @param scb
+	 * @throws IllegalArgumentException
+	 */
+	private void compressToTZip(final File file, final boolean force, final UpdResultCallBack cb, final UpdSrcCallBack scb) throws IllegalArgumentException
+	{
+		if("zip".equals(FilenameUtils.getExtension(file.getName())))
+			zip2TZip(file, force, cb);
+		else
+			Optional.ofNullable(sevenZip2Zip(file, true, cb, scb)).filter(File::exists).ifPresent(f -> zip2TZip(f, force, cb));
+	}
+
+	/**
+	 * @param file
+	 * @param force
+	 * @param cb
+	 * @param scb
+	 * @throws IllegalArgumentException
+	 */
+	private void compressToZip(final File file, final boolean force, final UpdResultCallBack cb, final UpdSrcCallBack scb) throws IllegalArgumentException
+	{
+		if("zip".equals(FilenameUtils.getExtension(file.getName())))
+		{
+			if (force)
+				zip2Zip(file, cb, scb);
+			else
+				cb.apply("Skipped");
+		}
+		else
+			sevenZip2Zip(file, false, cb, scb);
+	}
+
+	/**
+	 * @param file
+	 * @param force
+	 * @param cb
+	 * @param scb
+	 * @throws IllegalArgumentException
+	 */
+	private void compressToSevenZip(final File file, final boolean force, final UpdResultCallBack cb, final UpdSrcCallBack scb) throws IllegalArgumentException
+	{
+		switch (FilenameUtils.getExtension(file.getName()))
+		{
+			case "zip":
+				zip2SevenZip(file, cb, scb);
+				break;
+			case "7z":
+				if (force)
+					sevenZip2SevenZip(file, cb, scb);
+				else
+					cb.apply("Skipped");
+				break;
+			default:
+				sevenZip2SevenZip(file, cb, scb);
+				break;
 		}
 	}
 }
